Make wxCocoa compile in 64-bit mode.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@47182 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/cocoa/dc.mm
@@ -43,7 +43,7 @@ wxCocoaDCStack wxDC::sm_cocoaDCStack;
 inline void CocoaSetPenForNSBezierPath(wxPen &pen, NSBezierPath *bezpath)
 {
     [pen.GetNSColor() set];
-    const float *pattern;
+    const CGFloat *pattern;
     [bezpath setLineDash:pattern count:pen.GetCocoaLineDash(&pattern) phase:0.0];
     [bezpath setLineWidth:pen.GetWidth()];
     switch(pen.GetJoin())
